+++ src/index.ts
@@ -1,3 +1,4 @@
+export * from "./alloc.js";
 export * from "./byte-queue.js";
 export * from "./bytes.js";
 export * from "./deserializer.js";
@@ -6,5 +7,4 @@ export * from "./encoding/base64.js";
 export * from "./encoding/hex.js";
 export * from "./endianess.js";
 export * from "./serializer.js";
-export { setupSharedBuffer } from "./shared-buffers.js";
 export * from "./stream-deserializer.js";